refactor: reorganize common components and update component naming
- Move TheCheckbox, TheCounter, TheInput to common directory
- Rename MultiSelectEnhanced to TheMultiSelect and move to common directory
- Update all component imports across the application
- Update Footer description to match project branding
- Update test imports and remove snapshot assertion
