+++ packages/apps/board/src/components.d.ts
@@ -33,7 +33,6 @@ declare module 'vue' {
     HeatMapTooltip: typeof import('./components/board/HeatMapTooltip.vue')['default']
     Modal: typeof import('./components/board/Modal.vue')['default']
     ModalMenu: typeof import('./components/board/ModalMenu.vue')['default']
-    MultiSelectEnhanced: typeof import('./components/common/MultiSelectEnhanced.vue')['default']
     NavBar: typeof import('./components/NavBar.vue')['default']
     OptionsModal: typeof import('./components/board/OptionsModal.vue')['default']
     ProblemBlock: typeof import('./components/board/ProblemBlock.vue')['default']
@@ -64,9 +63,10 @@ declare module 'vue' {
     TeamInfoModal: typeof import('./components/board/TeamInfoModal.vue')['default']
     TeamProblemBlock: typeof import('./components/board/TeamProblemBlock.vue')['default']
     TeamUI: typeof import('./components/board/TeamUI.vue')['default']
-    TheCheckbox: typeof import('./components/TheCheckbox.vue')['default']
-    TheCounter: typeof import('./components/TheCounter.vue')['default']
-    TheInput: typeof import('./components/TheInput.vue')['default']
+    TheCheckbox: typeof import('./components/common/TheCheckbox.vue')['default']
+    TheCounter: typeof import('./components/common/TheCounter.vue')['default']
+    TheInput: typeof import('./components/common/TheInput.vue')['default']
+    TheMultiSelect: typeof import('./components/common/TheMultiSelect.vue')['default']
     Tooltip: typeof import('./components/flowbite/Tooltip.vue')['default']
     Utility: typeof import('./components/board/Utility.vue')['default']
   }
